chore: move components dir
